+++ app/routes/components/Popover/changelog.codegen.ts
@@ -15,4 +15,5 @@ export const changelog = ''
   + '- 2025-10-03: Fix missing dependencies in useEffect hooks.\n'
   + '- 2025-10-26: Using `Modal` component in `Popover.Portal` to handle the portal\n'
   + '  rendering.\n'
+  + "- 2025-10-27: Make `Popover.Trigger`'s render can have prop `isOpen`.\n"
   ;