+++ src/components/Popover/index.ts
@@ -1,3 +1,3 @@
-// Popover from copy-ui @ 2025-10-26.
+// Popover from copy-ui @ 2025-10-27.
 
 export { type Placement, Popover } from "./popover";